requirements.txt
gunicorn
flask
requests
slackclient
slackeventsapi
#numpy
#pandas
#spacy
#https://github.com/explosion/spacy-models/releases/download/en_core_web_md-2.3.1/en_core_web_md-2.3.1.tar.gz#egg=en_core_web_md